Refactored YearnDeposit into a folder and reducer into its own file
